feat: shoot REAL ho tro integer va so thuc tu PLC
Doi schema shoot sang REAL, parseShoot khi push data, them script migrate bang may cu.

+++ backend/controllers/machinesController.js
@@ -99,7 +99,7 @@ export const addMachine = async (req, res) => {
         nr INTEGER,
         machine_id VARCHAR(255),
         timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-        shoot INTEGER,
+        shoot REAL,
         cycle REAL,
         time_on REAL,
         time_off REAL,
